can successfully read, field notes has one attribute run_list
